move pool status cells
