+++ apps/web/src/components/PoolStatusTable/cells.tsx
@@ -1,7 +1,16 @@
-import { PoolCandidate, User } from "@gc-digital-talent/graphql";
+import { IntlShape } from "react-intl";
 
-import ChangeStatusDialog from "../../pages/Users/UserInformationPage/components/ChangeStatusDialog";
-import ChangeDateDialog from "../../pages/Users/UserInformationPage/components/ChangeDateDialog";
+import {
+  LocalizedString,
+  Maybe,
+  PoolCandidate,
+  User,
+} from "@gc-digital-talent/graphql";
+import { Link } from "@gc-digital-talent/ui";
+import { getLocalizedName } from "@gc-digital-talent/i18n";
+
+import ChangeDateDialog from "~/components/CandidateDialog/ChangeDateDialog";
+import ChangeStatusDialog from "~/components/CandidateDialog/ChangeStatusDialog";
 
 export const statusCell = (candidate: PoolCandidate, user: User) => (
   <ChangeStatusDialog selectedCandidate={candidate} user={user} />
@@ -10,3 +19,24 @@ export const statusCell = (candidate: PoolCandidate, user: User) => (
 export const expiryCell = (candidate: PoolCandidate, user: User) => (
   <ChangeDateDialog selectedCandidate={candidate} user={user} />
 );
+
+export function viewTeamLinkCell(
+  url: Maybe<string> | undefined,
+  displayName: Maybe<LocalizedString> | undefined,
+  intl: IntlShape,
+) {
+  return url ? (
+    <Link color="black" href={url}>
+      {intl.formatMessage(
+        {
+          defaultMessage: "<hidden>View team: </hidden>{teamName}",
+          id: "ActH9H",
+          description: "Text for a link to the Team table",
+        },
+        {
+          teamName: getLocalizedName(displayName, intl),
+        },
+      )}
+    </Link>
+  ) : null;
+}
